Create Best time to buy and sell stock 8.JAVA
You are given an array of prices where prices[i] is the price of a given stock on an ith day.
You want to maximize your profit by choosing a single day to buy one stock and choosing a different day to sell that stock in the future.
Return the maximum profit you can achieve from this transaction. If you cannot accomplish any profit, return 0.

Input Format

The first line of each test case or query contains an integer 'N' representing the size of the first array/list.
The second line contains 'N' single space-separated integers representing the elements in the array/list.

Constraints

No

Output Format

Integer representing the answer of the problem statement.