Create Tower Of Hanoi.JAVA
Tower of Hanoi is a mathematical puzzle where we have three rods and n disks. The objective of the puzzle is to move all disks from the source rod to the destination rod using the third rod (say auxiliary). The rules are :
1) Only one disk can be moved at a time.
2) A disk can be moved only if it is on the top of a rod.
3) No disk can be placed on the top of a smaller disk.
Print the steps required to move n disks from the source rod to the destination rod.
The source Rod is named as 'a, the auxiliary rod as 'b', and the destination rod as 'c'.


Input Format:
Integer N

Constraints:
0 <= n <= 20

Output Format:
Steps in different lines (in one line print source and destination rod name separated by space)

Sample Input 0:
2

Sample Output 0:
a b
a c
b c