Create Binary Search 61.JAVA
You have been given a sorted(in ascending order) integer array/list(ARR) of size N and an element X. Write a function to search this element in the given input array/list using 'Binary Search'. Return the index of the element in the input array/list. If the element is not present in the array/list, then return -1.

Input Format:
Value representing the size of the array
The line consists of 'n' space-separated integers representing the values present in the Array
Target Element which we need to find

Constraints:
No

Output Format:
Index of that target element

Sample Input 0:
3
1 2 3
0

Sample Output 0:
-1